django_crm/django_crm/requirements/base.txt
# base requirements
django>=4.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg[binary]
